requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
moviepy==1.0.3
gTTS==2.3.2
google-generativeai
python-dotenv==1.0.0
gunicorn==20.1.0
click